feat(routes): add create and index pages for route management
- Implemented a new page for creating routes with form validation and error handling.
- Added an index page to list existing routes with filtering, pagination, and actions for enabling/disabling and deleting routes.
- Introduced utility functions for OAuth redirect handling and validation of URLs.
- Created API endpoints for setting authentication cookies and managing OAuth configurations.
- Added unit tests for new utility functions and OAuth configuration validation.
