+++ app/pages/settings/routes/index.vue
@@ -1,0 +1,482 @@
+<script setup lang="ts">
+const notify = useNotify();
+const page = ref(1);
+const pageLimit = 9;
+const route = useRoute();
+const router = useRouter();
+const tableName = "route_definition";
+const { confirm } = useConfirm();
+const { getIncludeFields } = useSchema(tableName);
+const { createEmptyFilter, buildQuery, hasActiveFilters } = useFilterQuery();
+const { createLoader } = useLoader();
+const { isTablet } = useScreen();
+const { isMounted } = useMounted();
+const { registerPageHeader } = usePageHeaderRegistry();
+
+const pageIconColor = 'primary';
+
+registerPageHeader({
+  title: "Route Manager",
+  gradient: "cyan",
+});
+
+const { getId } = useDatabase();
+
+const showFilterDrawer = ref(false);
+const currentFilter = ref(createEmptyFilter());
+const showSystem = ref(route.query.system === 'true');
+const showCollectionRoutes = ref(route.query.collectionRoutes === 'true');
+
+watch(() => route.query.system, (v) => { showSystem.value = v === 'true' })
+watch(() => route.query.collectionRoutes, (v) => { showCollectionRoutes.value = v === 'true' })
+
+watch(showSystem, (v) => {
+  if ((route.query.system === 'true') !== v) {
+    const query = { ...route.query }
+    if (v) query.system = 'true'
+    else delete query.system
+    router.replace({ query })
+  }
+})
+watch(showCollectionRoutes, (v) => {
+  if ((route.query.collectionRoutes === 'true') !== v) {
+    const query = { ...route.query }
+    if (v) query.collectionRoutes = 'true'
+    else delete query.collectionRoutes
+    router.replace({ query })
+  }
+})
+
+const filterLabel = computed(() => {
+  const activeCount = currentFilter.value.conditions.length;
+  return activeCount > 0 ? `Filters (${activeCount})` : "Filter";
+});
+
+const filterVariant = computed(() => {
+  return hasActiveFilters(currentFilter.value) ? "solid" : "outline";
+});
+
+const filterColor = computed(() => {
+  return hasActiveFilters(currentFilter.value) ? "secondary" : "neutral";
+});
+
+const {
+  data: apiData,
+  pending: loading,
+  execute: fetchRoutes,
+} = useApi(() => "/route_definition", {
+  query: computed(() => {
+    const conditions: any[] = [];
+    if (!showSystem.value) {
+      conditions.push({ isSystem: { _eq: false } });
+    }
+    if (!showCollectionRoutes.value) {
+      conditions.push({ mainTable: { _is_null: true } });
+    }
+    const filterQuery = hasActiveFilters(currentFilter.value)
+      ? buildQuery(currentFilter.value)
+      : null;
+    if (filterQuery) {
+      conditions.push(filterQuery);
+    }
+
+    return {
+      fields: getIncludeFields(),
+      sort: "-createdAt",
+      meta: "*",
+      page: page.value,
+      limit: pageLimit,
+      ...(conditions.length > 0 && { filter: { _and: conditions } }),
+    };
+  }),
+  errorContext: "Fetch Routes",
+});
+
+const routesData = computed(() => apiData.value?.data || []);
+const total = computed(() => {
+  const meta = apiData.value?.meta
+  if (showSystem.value && showCollectionRoutes.value && !hasActiveFilters(currentFilter.value)) {
+    return meta?.totalCount ?? 0
+  }
+  return meta?.filterCount ?? 0
+});
+
+useSubHeaderActionRegistry([
+  {
+    id: "toggle-system-routes",
+    icon: "lucide:shield",
+    get label() {
+      return showSystem.value ? "Hide System" : "Show System";
+    },
+    get variant() {
+      return showSystem.value ? "solid" as const : "outline" as const;
+    },
+    get color() {
+      return showSystem.value ? "warning" as const : "neutral" as const;
+    },
+    size: "md",
+    side: "right",
+    order: 0,
+    onClick: () => {
+      showSystem.value = !showSystem.value;
+      page.value = 1;
+      fetchRoutes();
+    },
+  },
+  {
+    id: "toggle-collection-routes",
+    icon: "lucide:table",
+    get label() {
+      return showCollectionRoutes.value ? "Hide Collection Routes" : "Collection Routes";
+    },
+    get variant() {
+      return showCollectionRoutes.value ? "solid" as const : "outline" as const;
+    },
+    get color() {
+      return showCollectionRoutes.value ? "info" as const : "neutral" as const;
+    },
+    size: "md",
+    side: "right",
+    order: 1,
+    onClick: () => {
+      showCollectionRoutes.value = !showCollectionRoutes.value;
+      page.value = 1;
+      fetchRoutes();
+    },
+  },
+]);
+
+useHeaderActionRegistry([
+  {
+    id: "filter-routes",
+    icon: "lucide:filter",
+    get label() {
+      return filterLabel.value;
+    },
+    get variant() {
+      return filterVariant.value;
+    },
+    get color() {
+      return filterColor.value;
+    },
+    size: "md",
+    onClick: () => {
+      showFilterDrawer.value = true;
+    },
+    permission: {
+      and: [
+        {
+          route: "/route_definition",
+          actions: ["read"],
+        },
+      ],
+    },
+  },
+  {
+    id: "create-route",
+    label: "Create Route",
+    icon: "lucide:plus",
+    variant: "solid",
+    color: "primary",
+    size: "md",
+    to: "/settings/routes/create",
+    permission: {
+      and: [
+        {
+          route: "/route_definition",
+          actions: ["create"],
+        },
+      ],
+    },
+  },
+]);
+
+async function handleFilterApply(filter: FilterGroup) {
+  currentFilter.value = filter;
+  
+  if (page.value === 1) {
+    
+    await fetchRoutes();
+  } else {
+    
+    const newQuery = { ...route.query };
+    delete newQuery.page;
+    
+    await router.replace({
+      query: newQuery,
+    });
+  }
+}
+
+watch(
+  () => route.query.page,
+  async (newVal) => {
+    page.value = newVal ? Number(newVal) : 1;
+    await fetchRoutes();
+  },
+  { immediate: true }
+);
+
+const { execute: updateRouteApi, error: updateError } = useApi(
+  () => `/route_definition`,
+  {
+    method: "patch",
+    errorContext: "Toggle Route",
+  }
+);
+
+const { execute: deleteRouteApi, error: deleteError } = useApi(
+  () => `/route_definition`,
+  {
+    method: "delete",
+    errorContext: "Delete Route",
+  }
+);
+
+const routeLoaders = ref<Record<string, any>>({});
+
+const methodColorMap: Record<string, string> = {
+  GET: 'info',
+  POST: 'success',
+  PATCH: 'warning',
+  DELETE: 'error',
+};
+
+function getPublishedMethodsStat(routeItem: any) {
+  const availableSet = new Set(
+    (routeItem.availableMethods || [])
+      .filter((m: any) => m?.method)
+      .map((m: any) => m.method)
+  );
+  const filtered = (routeItem.publishedMethods || [])
+    .filter((m: any) => m?.method && availableSet.has(m.method));
+  return {
+    component: filtered.length ? 'UBadge' : undefined,
+    values: filtered.length ? filtered.map((m: any) => ({
+      value: m.method.toUpperCase(),
+      props: { color: methodColorMap[m.method] }
+    })) : undefined,
+    value: filtered.length ? undefined : '-',
+  };
+}
+
+function getRouteLoader(routeId: string) {
+  if (!routeLoaders.value[routeId]) {
+    routeLoaders.value[routeId] = createLoader();
+  }
+  return routeLoaders.value[routeId];
+}
+
+function getRouteHeaderActions(routeItem: any) {
+  if (routeItem.isSystem) {
+    return [];
+  }
+
+  const hasAssociatedTable = getId(routeItem.mainTable);
+  const { schemas } = useSchema();
+  const tableExists = hasAssociatedTable && Object.values(schemas.value).some(
+    (table: any) => getId(table) === getId(routeItem.mainTable)
+  );
+
+  return [
+    {
+      component: 'USwitch',
+      props: {
+        'model-value': routeItem.isEnabled,
+        disabled: getRouteLoader(getId(routeItem)).isLoading || tableExists
+      },
+      onClick: (e?: Event) => e?.stopPropagation(),
+      onUpdate: () => toggleEnabled(routeItem),
+      tooltip: tableExists ? 'Cannot disable route with associated table' : undefined
+    }
+  ];
+}
+
+function getRouteFooterActions(routeItem: any) {
+  
+  const hasAssociatedTable = getId(routeItem.mainTable);
+  const { schemas } = useSchema();
+  const tableExists = hasAssociatedTable && Object.values(schemas.value).some(
+    (table: any) => getId(table) === getId(routeItem.mainTable)
+  );
+
+  return [
+    {
+      label: 'Delete',
+      props: {
+        icon: 'i-lucide-trash-2',
+        variant: 'solid',
+        color: 'error',
+        size: 'sm',
+      },
+      disabled: routeItem.isSystem || tableExists,
+      onClick: (e?: Event) => {
+        e?.stopPropagation();
+        deleteRoute(routeItem);
+      },
+    }
+  ];
+}
+
+async function toggleEnabled(routeItem: any) {
+  
+  const newEnabled = !routeItem.isEnabled;
+
+  if (apiData.value?.data) {
+    const routeIndex = apiData.value.data.findIndex(
+      (r: any) => getId(r) === getId(routeItem)
+    );
+    if (routeIndex !== -1) {
+      apiData.value.data[routeIndex].isEnabled = newEnabled;
+    }
+  }
+
+  await updateRouteApi({ id: getId(routeItem), body: { isEnabled: newEnabled } });
+
+  if (updateError.value) {
+    
+    if (apiData.value?.data) {
+      const routeIndex = apiData.value.data.findIndex(
+        (r: any) => getId(r) === getId(routeItem)
+      );
+      if (routeIndex !== -1) {
+        apiData.value.data[routeIndex].isEnabled = !newEnabled;
+      }
+    }
+    return;
+  }
+
+  const { loadRoutes } = useRoutes();
+  const { registerDataMenuItems } = useMenuRegistry();
+  const { schemas } = useSchema();
+
+  await loadRoutes();
+  await registerDataMenuItems(Object.values(schemas.value));
+
+  notify.success("Success", `Route ${newEnabled ? "enabled" : "disabled"} successfully`);
+}
+
+async function deleteRoute(routeItem: any) {
+  const isConfirmed = await confirm({
+    title: "Delete Route",
+    content: `Are you sure you want to delete route "${routeItem.path}"? This action cannot be undone.`,
+    confirmText: "Delete",
+    cancelText: "Cancel",
+  });
+
+  if (isConfirmed) {
+    await deleteRouteApi({ id: getId(routeItem) });
+
+    if (deleteError.value) {
+      return;
+    }
+
+    await fetchRoutes();
+
+    const { loadRoutes } = useRoutes();
+    const { registerDataMenuItems } = useMenuRegistry();
+    const { schemas } = useSchema();
+
+    await loadRoutes();
+    await registerDataMenuItems(Object.values(schemas.value));
+
+    notify.success("Success", `Route "${routeItem.path}" has been deleted successfully!`);
+  }
+}
+
+</script>
+
+<template>
+  <div class="space-y-6">
+    <Transition name="loading-fade" mode="out-in">
+      <div v-if="loading || !isMounted">
+        <CommonLoadingState
+          title="Loading routes..."
+          description="Fetching routing configuration"
+          size="sm"
+          type="card"
+          context="page"
+        />
+      </div>
+
+      <div v-else class="space-y-6">
+        <div v-if="routesData.length" class="space-y-6">
+          <div
+            class="grid gap-4"
+            :class="
+              isTablet
+                ? 'grid-cols-2'
+                : 'grid-cols-1 md:grid-cols-2 lg:grid-cols-2 2xl:grid-cols-3'
+            "
+          >
+            <CommonSettingsCard
+              v-for="routeItem in routesData"
+              :key="getId(routeItem)"
+              :title="routeItem.path"
+              :description="routeItem.mainTable?.name"
+              :icon="routeItem.icon || 'lucide:circle'"
+              :icon-color="pageIconColor"
+              :card-class="'cursor-pointer transition-all'"
+              @click="navigateTo(`/settings/routes/${getId(routeItem)}`)"
+              :top-badge="routeItem.isSystem ? { label: 'System', color: 'info' } : undefined"
+              :stats="[
+                {
+                  label: 'Status',
+                  component: 'UBadge',
+                  props: {
+                    variant: 'soft',
+                    color: routeItem.isEnabled ? 'success' : 'warning',
+                  },
+                  value: routeItem.isEnabled ? 'Enabled' : 'Disabled'
+                },
+                {
+                  label: 'Published Methods',
+                  ...getPublishedMethodsStat(routeItem)
+                }
+              ]"
+              :actions="getRouteFooterActions(routeItem)"
+              :header-actions="getRouteHeaderActions(routeItem)"
+            </CommonSettingsCard>
+          </div>
+        </div>
+
+        <CommonEmptyState
+          v-else-if="!loading"
+          title="No routes found"
+          description="No routing configurations have been created yet"
+          icon="lucide:route"
+          size="sm"
+        />
+
+        <div
+          v-if="!loading && routesData.length > 0 && total > pageLimit"
+          class="flex items-center justify-between mt-6"
+        >
+          <UPagination
+            v-model:page="page"
+            :items-per-page="pageLimit"
+            :total="total"
+            show-edges
+            :sibling-count="1"
+            :to="
+              (p) => ({
+                path: route.path,
+                query: { ...route.query, page: p },
+              })
+            "
+          />
+          <p class="hidden md:block text-sm text-[var(--text-quaternary)]">
+            Showing <span class="text-[var(--text-secondary)]">{{ (page - 1) * pageLimit + 1 }}-{{ Math.min(page * pageLimit, total) }}</span> of <span class="text-[var(--text-secondary)]">{{ total }}</span> results
+          </p>
+        </div>
+      </div>
+    </Transition>
+  </div>
+
+  <FilterDrawerLazy
+    v-model="showFilterDrawer"
+    :table-name="tableName"
+    :current-filter="currentFilter"
+    @apply="handleFilterApply"
+  />
+</template>
