+++ app/utils/auth.ts
@@ -1,0 +1,14 @@
+export function resolveOAuthRedirectTarget(
+  redirect: string | undefined,
+  currentUrl: string,
+) {
+  try {
+    if (redirect && redirect.length > 0) {
+      return new URL(redirect, currentUrl).toString();
+    }
+
+    return currentUrl;
+  } catch {
+    return currentUrl;
+  }
+}
